Add energy source estimates
This is a persisted model of calculated data.
For an internal meter (measuring local energy usage)
we can try to estimate how much energy was sourced from
local generators and how much energy was sourced from the
grid.

This should be helpful in judging whether energy was used
at the "right" time (i.e. when generation was sufficient).

I am also planning to use it, to determine the cost of electricity
for an internal meter more precisely.