Consider export during energy source estimation
This was a flaw in my previous calculation.
An export measurement essentially means, that the energy
generated wasn't available for the internal consumer.

Thus we have to deduct the export from the net generation.

Conversely this will also affect the "total energy" part
of the fraction, but this is in line with the calculation
of site local consumption, which is also expressed as

    import + generation - export

Due to resolution issues we have to truncate fractions
in case export is bigger than generation.
This might still lead to generation being over-estimated:

E.g. A low resolution export alternating between two values A and B
should be expected to effectively export at the average of A and B.
But instead, truncation at the higher export value will cause the
"above average" fraction of the export to remain unconsidered,
while fully counting generation during the lower export period.