Allow to change charging efficiency of battery
This should give slightly better estimations of
battery performance, since some energy is just converted
to heat, but not charged, even though it is consumed.

Technically there is most likely a charging and a discharging
efficiency, though so far I only found specs that gave the round-
trip efficiency, so this will have to do as an approximation.

This simplified approach means, that we expect to discharge the full
capacity into usable energy, even though this will most likely not
happen.