Allow to consider resets of meters
For now this enables a manual process where you can enter
reset from and to values, e.g. when a physical meter is swapped
for another meter.
This allows to keep the same "meter" entitiy in the application
and follow a continuous usage graph, even when the physical devise
measuring has been replaced.

Yet missing:
* Allowing to re-combine meters
* Automatically detecting this condition for meters that auto-reset